updating unittests
